Break gnome-shell's circular build dep on itself
... by getting rid of the gnome-bluetooth buildrequires.

gnome-shell was dragging in an older version of gnome-shell at build
time through the following dep chain: gnome-shell BRs gnome-bluetooth,
which in turn requires desktop-notification-daemon, which is a virtual
package provided by gnome-shell.

We need to break the circular build dep in order to rebuild gnome-shell
with new libcogl. If it turns out gnome-bluetooth (as opposed to
gnome-bluetooth-libs-devel) is really needed at build time, then this
commit can be reverted after the libcogl related rebuild is done.