Use %autopatch instead of doing our own custom git-am trick
There's some quirks here. We have a .gitignore we'd like to use, so we
need to copy that in before doing git-init. But the git-init happens in
%autosetup if you pass it -S git, so we can't do that. Also we want to
blow away any .git directory inside the tarball, otherwise the initial
commit will fail. So, unpack, nuke .git, copy in .gitignore, then
manually do the rest of %autosetup.