mutter: reduce 'dbus-x11' dependency to 'dbus'
Back in the days a dependency on 'dbus-x11' was required to
automatically start a session bus when Xorg starts up. This is no
longer the case, as the 'dbus' package now always provides a system-
and user-bus to its dependents (currently via systemd units).

Lets reduce the 'dbus-x11' dependency to 'dbus', so we no longer
explicitly pull in 'dbus-daemon' recursively, but can safely support
other bus providers that don't support the legacy X11 startup scripts
(in particular dbus-broker).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>